# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lasolution API
  description: |
    Contrat HTTP v1 — JWT Bearer (`accessToken` depuis `/auth/login` si PostgreSQL actif),
    commandes/colis, webhooks PSP async (worker), paiements multi-rail, stubs Zoho/transporteur.
    Schémas alignés sur `backend/prisma/schema.prisma`.
  version: 0.2.0
servers:
  - url: http://localhost:4000
    description: Développement local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton `accessToken` retourné par POST /auth/login (nécessite DB + `API_JWT_SECRET`).
    SessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session NextAuth (front) — distinct du Bearer API.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      required: [id, email, role, name]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, client, relais, solupacker, solu_livreur, ambassadeur]
        name:
          type: string
        accessToken:
          type: string
          description: Présent si la base est joignable (JWT court).
        refreshToken:
          type: string
          description: Rotation via POST /auth/refresh.
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
paths:
...
  /auth/login:
    post:
      tags: [auth, admin, client, relais, packer, livreur, ambassadeur]
      summary: Connexion (démo credentials / mots de passe d’environnement)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Utilisateur authentifié (profil minimal)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit connexion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotation du refresh token (révoque l’ancien, en émet un nouveau)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Nouveaux jetons
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, refreshToken, user]
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    type: object
                    required: [id, email, role, name]
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      role:
                        type: string
                      name:
                        type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Sans base de données
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...